web_assistant/backend/requirements.txt
# Core Django
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Web Scraping
playwright==1.41.2
scrapy==2.11.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# Task Queue
celery==5.3.6

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.23.2

# Code Quality
black==23.12.1
isort==5.13.2
mypy==1.7.1
